requirements.txt
backoff==2.2.1
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.1.0
click==8.1.3
Deprecated==1.2.13
googleapis-common-protos==1.59.0
grpcio==1.54.0
idna==3.4
importlib-metadata==6.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp-proto-grpc==1.25.0
opentelemetry-exporter-otlp-proto-http==1.25.0
opentelemetry-instrumentation==0.46b0
opentelemetry-instrumentation-logging==0.46b0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
opentelemetry-util-http==0.46b0
protobuf==4.22.4
requests==2.30.0
typing_extensions==4.5.0
urllib3==1.26.15
wrapt==1.15.0
zipp==3.15.0
eclipse-zenoh==0.10.1rc0
pycdr2==1.0.0